Converted price and quantity to BigDecimal for more accuracy

+++ src/main/kotlin/com/example/lob/limitorderbook/LimitPriceOrders.kt
@@ -1,11 +1,12 @@
 package com.example.lob.limitorderbook
 
 import com.example.lob.order.Order
+import java.math.BigDecimal
 import java.util.*
 
 data class LimitPriceOrders(
-    val price: Double,
-    var quantity: Double,
+    val price: BigDecimal,
+    var quantity: BigDecimal,
     val orders: LinkedList<Order>,
     val orderCount: Int = orders.size
 )
