+++ src/main/kotlin/com/example/lob/order/OrderResponseDto.kt
@@ -1,11 +1,12 @@
 package com.example.lob.order
 
 import com.example.lob.currencypair.CurrencyPair
+import java.math.BigDecimal
 
 data class OrderResponseDto(
     val side: BuyOrSellEnum,
-    val quantity: Double,
-    val price: Double,
+    val quantity: BigDecimal,
+    val price: BigDecimal,
     val currencyPair: CurrencyPair,
     val orderCount: Int
 )
